Upgrade to 3.0.9.
Add patches provided by upstream to deal with broken build due to swig.
Switch to linux-api-headers in makedepends, as it is providing /usr/include/linux/audit.h for the build.
Use local array to simplify the use of configure options in build().
Pick files for python-audit in audit's build() instead of running make twice and deleting files.